+++ examples/6-assets/gltf-test/src/lib.rs
@@ -58,7 +58,7 @@ static mut ANIMATION: u32 = 0;
 
 // Animation state
 static mut FRAME_COUNT: u16 = 0;
-static mut BONE_COUNT: u8 = 0;
+static mut BONE_COUNT: u16 = 0;
 static mut ANIM_TIME: f32 = 0.0;
 static mut ANIM_SPEED: f32 = 1.0;
 static mut PAUSED: bool = false;
@@ -97,8 +97,8 @@ pub extern "C" fn init() {
         ANIMATION = rom_keyframes(b"test_anim".as_ptr(), 9);
 
         // Query animation properties
-        FRAME_COUNT = keyframes_frame_count(ANIMATION);
-        BONE_COUNT = keyframes_bone_count(ANIMATION);
+        FRAME_COUNT = keyframes_frame_count(ANIMATION) as u16;
+        BONE_COUNT = keyframes_bone_count(ANIMATION) as u16;
     }
 }
